Stream Enrich: bump SBT to 0.13.16 (closes #3382)

+++ 3-enrich/stream-enrich/src/main/scala/com.snowplowanalytics.snowplow.enrich/stream/sources/AbstractSource.scala
@@ -115,7 +115,7 @@ object AbstractSource {
    * @param evt
    * @return size
    */
-  def getSize(evt: String): Long = ByteBuffer.wrap(evt.getBytes(UTF_8)).capacity
+  def getSize(evt: String): Long = ByteBuffer.wrap(evt.getBytes(UTF_8)).capacity.toLong
 }
 
 /**
